requirements.txt
fastapi[standard]
pydantic-settings
asyncpg
python-dotenv
SQLAlchemy
alembic
cloudinary
passlib[argon2]
python-jose[cryptography]
requests